requirements.txt
pymupdf
sentence-transformers
faiss-cpu
groq
mcp